chore: optimize code structure
